URL 접속 후 조회 데이터 검증 테스트 화면 수정

+++ src/main/java/com/zipdoc/dev/controller/rest/PasteController.java
@@ -79,12 +79,12 @@ public class PasteController {
      * @throws Exception
      */
     @GetMapping("update/{pasteSeqNo}")
-    public ResponseEntity<ResponseVo<Integer>> modifyPasteCount(HttpServletRequest request, @PathVariable("pasteSeqNo") Long pasteSeqNo)
+    public ResponseEntity<ResponseVo<Long>> modifyPasteCount(HttpServletRequest request, @PathVariable("pasteSeqNo") Long pasteSeqNo)
             throws Exception {
 
-        ResponseVo<Integer> resp = new ResponseVo<Integer>();
+        ResponseVo<Long> resp = new ResponseVo<Long>();
 
-        Integer result = pasteService.updatePasteCount(pasteSeqNo, request);
+        Long result = pasteService.updatePasteCount(pasteSeqNo, request);
 
         if (result == null) {
             return ResponseEntity.ok(new ResponseVo(RspCode.INTERNAL_SERVER));
